Refactoring main page and pizzas components components
- Renamed stories directory to contexts, in this directory
  there are application's API contexts

- Added specific component to show pizza size details and
  abstract this piece of the main page

- Added directory faker-data to put all fakers data that will
  be used to do the views while doesn't have an API implementation
