+++ orders-system/src/pages/Main/index.js
@@ -1,11 +1,11 @@
 import React, { useContext } from 'react';
 
-import Pizza from 'components/Pizza';
-import { AuthContext } from 'stories/Auth';
-import pizzaSizes from 'services/Pizzas';
+import pizzaSizes from 'faker-data/pizzas';
+import { AuthContext } from 'contexts/Auth';
 import Content from 'templates/pages/Content';
-import { firstName, singularOrPlural } from 'helpers/formatter';
+import { firstName } from 'helpers/formatter';
 import DefaultPageLayout from 'templates/pages/Default';
+import PizzaSizeDetails from 'components/PizzaSizeDetails';
 
 import * as S from './style';
 
@@ -23,23 +23,9 @@ const Main = () => {
             Escolha o tamanho da pizza
           </S.TitlePizzas>
           <S.ContainerPizzaSizes>
-            { pizzaSizes.map(({
-              id,
-              name,
-              size,
-              slices,
-              flavours,
-            }) => (
-              <S.GridPizzaSize key={id}>
-                <S.PaperSizeDetails>
-                  <Pizza size={size} />
-                  <S.Divider />
-                  <S.SizeName>{ name }</S.SizeName>
-                  <S.SizeSlaces>
-                    { slices } { singularOrPlural(slices, 'fatia', 'fatias')}, {' '}
-                    { flavours } { singularOrPlural(flavours, 'sabor', 'sabores') }
-                  </S.SizeSlaces>
-                </S.PaperSizeDetails>
+            { pizzaSizes.map((pizzaSize) => (
+              <S.GridPizzaSize key={pizzaSize.id}>
+                <PizzaSizeDetails details={pizzaSize} />
               </S.GridPizzaSize>
             ))}
           </S.ContainerPizzaSizes>
