+++ orders-system/src/pages/Main/style.js
@@ -1,10 +1,6 @@
 import styled from 'styled-components';
 
-import {
-  Grid,
-  Typography,
-  Paper, Divider as MaterialDivider,
-} from '@material-ui/core';
+import { Grid, Typography } from '@material-ui/core';
 
 /**
  * Pizzas content container
@@ -40,34 +36,3 @@ export const GridPizzaSize = styled(Grid).attrs({
   item: true,
   xs: true,
 })``;
-
-/**
- * Paper to describe details about pizza size
- */
-export const PaperSizeDetails = styled(Paper)`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px 0;
-  min-width: 250px;
-`;
-
-/**
- * Text size pizza name
- */
-export const SizeName = styled(Typography).attrs({
-  variant: 'h5',
-})``;
-
-/**
- * Text about the size slaces
- */
-export const SizeSlaces = styled(Typography)``;
-
-/**
- * Divider details about pizza size
- */
-export const Divider = styled(MaterialDivider)`
-  margin: 20px 0;
-  width: 100%;
-`;
